+++ JsonSchemaAndStructs/KiteConnect-Java-Classes/TickerLtp/TickerLtp.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.*;
 
 public class TickerLtp {
     private Long instrumentToken;
-    private Long lastPrice;
+    private Double lastPrice;
     private String mode;
     private Boolean tradable;
 
@@ -14,9 +14,9 @@ public class TickerLtp {
     public void setInstrumentToken(Long value) { this.instrumentToken = value; }
 
     @JsonProperty("last_price")
-    public Long getLastPrice() { return lastPrice; }
+    public Double getLastPrice() { return lastPrice; }
     @JsonProperty("last_price")
-    public void setLastPrice(Long value) { this.lastPrice = value; }
+    public void setLastPrice(Double value) { this.lastPrice = value; }
 
     @JsonProperty("mode")
     public String getMode() { return mode; }
